readelf.c: fix a compiler warning
Fix the following compiler warning:

  CC     readelf.lo
readelf.c: In function ‘get_arch_from_elf’:
readelf.c:72: warning: comparison between signed and unsigned

The problem is when comparing signed and unsigned, both operands are
cast to unsigned, therefore the code was not able to detect that read()
returned -1. By casting result of sizeof() to int we are now comparing
signed integers and can now detect -1 from read().

+++ src/lib/readelf.c
@@ -67,7 +67,7 @@ int get_arch_from_elf(const char *file)
 		return -1;
 	nbytes = read(fd, (void *) &elf_hdr, sizeof(elf_hdr));
 	close(fd);
-	if (nbytes < sizeof(elf_hdr))
+	if (nbytes < (int)sizeof(elf_hdr))
 		return -1;
 	if (check_elf_magic(elf_hdr.ident))
 		return -1;
